+++ src/env-var/cc-env-var-form.smart-exposed-config.js
@@ -3,10 +3,11 @@ import '../smart/cc-smart-container.js';
 import { getAllExposedEnvVars, updateAllExposedEnvVars } from '@clevercloud/client/esm/api/v2/application.js';
 import { toNameValueObject } from '@clevercloud/client/esm/utils/env-vars.js';
 import { fetchApp } from '../lib/api-helpers.js';
+import { i18n } from '../lib/i18n.js';
+import { notifyError, notifySuccess } from '../lib/notifications.js';
 import { fromCustomEvent, LastPromise, merge, unsubscribeWithSignal, withLatestFrom } from '../lib/observables.js';
 import { sendToApi } from '../lib/send-to-api.js';
 import { defineComponent } from '../lib/smart-manager.js';
-import { addErrorType } from '../lib/utils.js';
 
 defineComponent({
   selector: 'cc-env-var-form[context="exposed-config"]',
@@ -24,13 +25,12 @@ defineComponent({
 
     const onSubmit$ = fromCustomEvent(component, 'cc-env-var-form:submit')
       .pipe(withLatestFrom(context$));
-    const onDismissedErrors$ = fromCustomEvent(component, 'cc-env-var-form:dismissed-error');
 
     unsubscribeWithSignal(disconnectSignal, [
 
       error$.subscribe(console.error),
-      error$.subscribe((error) => {
-        component.error = error.type;
+      error$.subscribe(() => {
+        component.error = true;
         component.saving = false;
       }),
       app_lp.value$.subscribe((app) => (component.appName = app.name)),
@@ -41,33 +41,32 @@ defineComponent({
 
       onSubmit$.subscribe(([variables, { apiConfig, ownerId, appId }]) => {
 
-        component.error = null;
+        component.error = false;
         component.saving = true;
 
-        variables_lp.push(() => {
-          return updateExposedConfig({ apiConfig, ownerId, appId, variables })
-            .then(() => variables)
-            .catch(addErrorType('saving'));
-        });
-      }),
-
-      onDismissedErrors$.subscribe(() => {
-        component.error = null;
-        component.saving = false;
+        updateExposedConfig({ apiConfig, ownerId, appId, variables })
+          .then(() => {
+            component.variables = variables;
+            notifySuccess(component, i18n('cc-env-var-form.update.success'));
+          })
+          .catch(() => notifyError(component, i18n('cc-env-var-form.update.error')))
+          .finally(() => {
+            component.saving = false;
+          });
       }),
 
       context$.subscribe(({ apiConfig, ownerId, appId }) => {
 
-        component.error = null;
+        component.error = false;
         component.saving = false;
         component.variables = null;
 
         if (apiConfig != null && ownerId != null && appId != null) {
           app_lp.push((signal) => {
-            return fetchApp({ apiConfig, signal, ownerId, appId }).catch(addErrorType('loading'));
+            return fetchApp({ apiConfig, signal, ownerId, appId });
           });
           variables_lp.push((signal) => {
-            return fetchExposedConfig({ apiConfig, signal, ownerId, appId }).catch(addErrorType('loading'));
+            return fetchExposedConfig({ apiConfig, signal, ownerId, appId });
           });
         }
       }),
@@ -86,5 +85,6 @@ function fetchExposedConfig ({ apiConfig, signal, ownerId, appId }) {
 
 function updateExposedConfig ({ apiConfig, signal, ownerId, appId, variables }) {
   const variablesObject = toNameValueObject(variables);
-  return updateAllExposedEnvVars({ id: ownerId, appId }, variablesObject).then(sendToApi({ apiConfig, signal }));
+  return updateAllExposedEnvVars({ id: ownerId, appId }, variablesObject)
+    .then(sendToApi({ apiConfig, signal }));
 }